requirements.txt
pillow== 9.0.1
numpy==1.22.4
scipy==1.8.1
matplotlib==3.5.3
Pillow==9.0.1
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.10.1